Add files via upload
These file will first record a 15 second video and store them with a file name "Video_Output.mp4". Then it will store the GPS location of the device based on the IP address as there is no GPS service on laptop. So due to which the location might not be accurate depending on the IP address. Then it will store the GPS location in the file "GPS_Location.txt". Then it will send the mail to specified email address with a message "Its an Emergency! I am attaching you the GPS location and a 15 seconds Video Respond ASAP!" and with a subject "CODE BLUE [ EMERGENCY (SOS) ]". It will also attach the files "Video_Output.mp4" and "GPS_Location.txt" as an attachment.